LPD-78234 Ignore reverse relationships since they cannot be removed

+++ modules/apps/site/site-cms-site-initializer/src/main/java/com/liferay/site/cms/site/initializer/internal/struts/UpdateStructureStrutsAction.java
@@ -114,8 +114,10 @@ public class UpdateStructureStrutsAction implements StrutsAction {
 						getObjectRelationshipsByObjectDefinitionId2(
 							objectDefinitionId, false)) {
 
-			_objectRelationshipLocalService.deleteObjectRelationship(
-				serviceBuilderObjectRelationship);
+			if (!serviceBuilderObjectRelationship.isReverse()) {
+				_objectRelationshipLocalService.deleteObjectRelationship(
+					serviceBuilderObjectRelationship);
+			}
 		}
 	}
